Changes to type system.
Superficial changes
->Renamed Traceable interface to IPath.
->Renamed Path interface to IPath2.

Substantial changes
->Created a class called Path, which implements IPath. So there is both
class and an interface for the concept of a Path. Why? Interfaces and
classes serve different purposes. Interfaces allow one to treat two
types as the same. Classes do this too. In addition, classes allow for
code reuse. A subclass can make use of a superclass's implementation of
a method. But there is a tradeoff to using classes. It makes types more
brittle, more resistent to change. So interfaces are more flexible than
classes but aren't as good for code reuse. By providng an interface, it
allows users of this library to create their own Paths that can belong
to a totally different class hierarchy. This wouldn't be possible if
Path was only available as a class. By providing a class, it allows me
to implement variables and methods that don't have to be redundantly
reimplemented.